// frontend/src/features/companies/types.ts
export interface CompanyPublic {
  id: string
  name: string
  logoUrl: string | null
  description: string | null
  companySize: string | null
  industry: string | null
  website: string | null
  contactEmail: string | null
  contactPhone: string | null
  address: string | null
  createdAt: string
}